Set up instant-setup, move store logic into lib/app
