+++ src/app/apps/[id]/metadata/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { getApp } from "@/app/actions";
+import { getApp } from "@/lib/app";
 import { appIdpEntityId, appIdpRedirectUrl, appLoginUrl } from "@/lib/app";
 import { INSECURE_PUBLIC_CERTIFICATE } from "@/lib/insecure-cert";
